# lib/api-zod/src/generated/types/caseDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Account } from './account';
import type { Audit } from './audit';
import type { Case } from './case';
import type { Complaint } from './complaint';
import type { FundFlow } from './fundFlow';
import type { Intervention } from './intervention';
import type { LastCredited } from './lastCredited';
import type { PredictionResult } from './predictionResult';
import type { Recommendation } from './recommendation';
import type { Risk } from './risk';
import type { Transaction } from './transaction';
import type { Vasp } from './vasp';
import type { Wallet } from './wallet';

export type CaseDetail = Case & {
  complaint: Complaint;
  accounts: Account[];
  transactions: Transaction[];
  fundFlow: FundFlow;
  wallets: Wallet[];
  vasp: Vasp[];
  risk: Risk;
  predictions: PredictionResult;
  recommendations: Recommendation[];
  lastCredited: LastCredited;
  intervention: Intervention;
  audit: Audit[];
};