# lib/api-zod/src/generated/types/complaintInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ComplaintInput {
  /** @minLength 10 */
  description: string;
  transactionIds?: string[];
  accounts?: string[];
  upiIds?: string[];
  wallets?: string[];
  incidentDate?: string;
  location?: string;
}