# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Case } from './case';
import type { DashboardAlertsItem } from './dashboardAlertsItem';
import type { DashboardMetrics } from './dashboardMetrics';
import type { DashboardRiskDistributionItem } from './dashboardRiskDistributionItem';
import type { DashboardTransactionVolumeItem } from './dashboardTransactionVolumeItem';

export interface Dashboard {
  metrics: DashboardMetrics;
  transactionVolume: DashboardTransactionVolumeItem[];
  riskDistribution: DashboardRiskDistributionItem[];
  recentCases: Case[];
  alerts: DashboardAlertsItem[];
  conversionWindow: string;
}