# lib/api-zod/src/generated/types/graphNode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GraphNode {
  id: string;
  label: string;
  kind: string;
  risk: number;
  x: number;
  y: number;
}