# lib/api-zod/src/generated/types/intervention.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Intervention {
  id: string;
  status: string;
  requestType: string;
  caseId: string;
  account: string;
  bank: string;
  branch: string;
  ifsc: string;
  reason: string;
  approvalRequired: boolean;
  /** @nullable */
  submittedAt?: string | null;
}