# lib/api-zod/src/generated/types/lastCredited.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LastCredited {
  account: string;
  transaction: string;
  amount: number;
  timestamp: string;
  risk: string;
  bank: string;
  branch: string;
  ifsc: string;
}