# lib/api-zod/src/generated/types/predictionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Hotspot } from './hotspot';

export interface PredictionResult {
  hotspots: Hotspot[];
  generatedAt: string;
  modelVersion: string;
}