# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Recommendation {
  priority: string;
  title: string;
  reason: string;
  evidence: string[];
  confidence: number;
}