# lib/api-zod/src/generated/types/risk.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Risk {
  score: number;
  category: string;
  confidence: number;
  features: string[];
  modelVersion: string;
}